+++ packages/react-form/src/index.ts
@@ -4,7 +4,7 @@ export * from "./form/components/form";
 // Types
 export * from "./form/types";
 
-// Cleaner API on core ui-shad components
+// Cleaner API on core shadcn components
 export * from "./components/checkbox-with-label";
 export * from "./components/select";
 export * from "./components/checkbox-group";