+++ packages/react-table/src/lib/feature-props/merge-feature-props.ts
@@ -1,4 +1,4 @@
-import { mergeStyleProps } from "@zoos/ui-shad";
+import { mergeStyleProps } from "@zoos/shadcn";
 import { type ComponentProps } from "./types";
 
 const mergeFeatureProps = (