+++ packages/shadcn/vite.config.ts
@@ -8,7 +8,7 @@ import { nxCopyAssetsPlugin } from "@nx/vite/plugins/nx-copy-assets.plugin";
 
 export default defineConfig({
   root: __dirname,
-  cacheDir: "../../node_modules/.vite/packages/ui-shad",
+  cacheDir: "../../node_modules/.vite/packages/shadcn",
   plugins: [
     react(),
     nxViteTsPaths(),
@@ -25,7 +25,7 @@ export default defineConfig({
   // Configuration for building your library.
   // See: https://vitejs.dev/guide/build.html#library-mode
   build: {
-    outDir: "../../dist/packages/ui-shad",
+    outDir: "../../dist/packages/shadcn",
     emptyOutDir: true,
     reportCompressedSize: true,
     commonjsOptions: {
@@ -34,7 +34,7 @@ export default defineConfig({
     lib: {
       // Could also be a dictionary or array of multiple entry points.
       entry: "src/index.ts",
-      name: "ui-shad",
+      name: "shadcn",
       fileName: "index",
       // Change this to the formats you want to support.
       // Don't forget to update your package.json as well.
@@ -52,7 +52,7 @@ export default defineConfig({
     include: ["src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}"],
     reporters: ["default"],
     coverage: {
-      reportsDirectory: "../../coverage/packages/ui-shad",
+      reportsDirectory: "../../coverage/packages/shadcn",
       provider: "v8",
     },
   },